CBL-6000: Lazy Vector Index Updater doesn't recongize N1QL dictionary and array literal expression
* We currently don't store the query language of the index expression in the database. When retrieved from the db, we used heruistics to determine whether it's JSON or N1QL. This heuristic broke down when the vector index is lazy and expression is N1QL, somehow like, "{'key': property}". The heuristic leads to N1QL expression.

To unqmbiguously distinguish between JSON and N1QL, we use illegal prefixes, "=j" for JSON and "=n" for N1QL. We prefix the expression when we register it with database table, and remove them when we retriieve it from db table, if they are present. If the prefix is present, we use the prefix to determine its QueryLanguage. If not present, we assume it is registered prior to this commit and we use the current heuristic.

The prefixed expressions are not parsable, and so they should not exist in the existing index registry, and they should not exist in objects of IndexSpec.