Fixed ambiguous use of `endian::`
If the C++20 header <bit> has been included, and the source file
is using both `std` and `fleece`, then `endian` becomes an
ambiguous identifier that can mean both the `std::endian` class and
the `fleece::endian` namespace.
I couldn't think of a better name for `fleece::endian`, and didn't
want to rename every use of it, so I just resolved the ambiguities.