Allow a Timer to destruct itself during its callback
Fixes a longstanding limitation that IIRC we knew about and just
avoided. I couldn't avoid it in the code I'm working on, so I finally
fixed it.

Manager::unschedule(), if the timer's being deleted, now checks to see
if it's being called on the Manager's thread. If so, it no longer
waits for the timer's `_triggered` flag to clear (which caused a
deadlock.)

Instead it checks if the timer is the same as the new
`_triggeredTimer` variable; if so, it clears the variable.

When control returns to the run() method, if `_triggeredTimer` has
been cleared it won't clear the `_triggered` flag or check the
`_autoDelete` flag.