Fixed NotFound error creating BlobStore
Discovered this scenario while testing replications in liteserv:
1. Create an empty database
2. Close it
3. Reopen it without giving the kC4DB_Create flag
4. Access the BlobStore, i.e. call c4Database::getBlobStore()

Result: getBlobStore fails with a NotFound error.
The reason is that step 1 didn't instantiate the BlobStore or create
the Attachments directory; and then in step 4 the db's flags don't
include kC4DB_Create so the BlobStore instance won't create the
directory.

Fixed by passing the kC4DB_Create flag to the C4BlobStore constructor
whenever the database is writeable.