Make sync work across VV upgrades (#1913)
* Expose DataFile::_log() as public

* Make sync work across VV upgrades

The transition from tree-based revids to version vectors is tricky
for replication, and the prior approach (adding a fake 'legacy'
version when creating the vector) didn't turn out to work.

Instead, when a doc is upgraded in memory to a VectorRecord it keeps
its tree-based/legacy revids. When the doc is saved it creates a
version vector for the local Revision but also remembers the prior
legacy revid. Remote Revisions keep their tree-based IDs until VVs
are pulled from or pushed to those remotes.

Revision histories may now include one or more legacy revids at the
end. These are what allow the recipient of a revision (an active or
passive puller) to match up an incoming version vector with a local
legacy revid.

There's still a question of how long we need to keep the legacy revid
of the document. It only needs to be used once per remote, but in a
P2P scenario you don't know if you might have to sync with some new
peer that hasn't upgraded that doc yet. In any case the revid isn't
very big (~24 bytes) so it may not be worth removing.

* Removed unnecessary VV-upgrade constants

Some constants added earlier for VV upgrading and dealing with legacy
revids, which have been superseded by newer techniques, were nuked:

- kDocGetUpgraded / kUpgrade
- kLegacyRevSourceID
- VectorRecord::_versioning

* Fixed some broken test code

c4doc_getRevisionHistory requires doc be loaded with kDocGetAll.
VectorDoc correctly threw an exception on this.
Also, all this is doing is checking the current rev, so no need to
get the history.