CBL-6239: Fixed c4doc_getRevisionBody in a 2.x db (#2136)
VectorRecord didn't properly manage the `body` column of a doc from a
2.x db that doesn't have the `extra` column. In such a db the body is
the encoded rev tree, not just the current rev. VectorRecord was
returning that entire slice from its currentRevisionData method.

I added a new member _bodyDocRange, a slice that's just the current
rev within the _bodyDoc. Normally it's the same as _bodyDoc.data(),
but in a 2.x doc it encompasses just the current rev body.

Fixes CBL-6239

Co-authored-by: Jianmin Zhao <jianmin.zhao@couchbase.com>