CBL-6255: fl_each with group-by not working
The cause is found to be the following. The virtual table generated by fl_each includes columns of "body" and "data". The former stores a pointer to preparsed value and the latter stores the original fleece data. When Group-By is called for, SQLite may regenerate rows and renders the body column stale. In order to keep the performance benefit of the body column, we use the data column only when the query uses the GROUP-BY clause.