CBL-6332: Implicit key change in query results (#2182)
In V3.2, when WHAT clause of the JSON query is absent of empty, we will assign meta properties, _id and _sequence, to the what clause, and use then property names verbatim in the corresponding column names (keep the underscore in the names). In this commit, we mean to keep this behavior with new query translator.
The column names are used in the query result of the platform API as the dictionary keys.