CBL-6285: Explain for predictive query no longer names an index
Update the default table name from main source in the query expression.

When we create the query object from the database, with, for example, SQLiteDataFile::compileQuery, the defaultTableName of QueryTranslator is initialized with the table name of the default collection. This table is assumed to be the main data source of the query expression. Therefore, we need to update it as we parse the first item in the FROM clause, before PredictionNode is created.