Ensure sequence indexes are created when DBAccess starts
The index is created lazily. If a collection doesn't have one, an
operation using a read-only db from the pool might trigger creating
it, which will fail because it's read-only.

To work around this, DBAccess's constructor uses the writeable db
to call call `getDocumentBySequence` on every collection, forcing the
index to be created.