Fixed DatabasePool deadlock
When Pusher::sendRevision() fails, it calls doneWithRev(). Both
methods were borrowing the db from the pool, so the second borrow
deadlocked.
Fixed this by having sendRevision release the db before calling
doneWithRev.

Also added diagnostics to DatabasePool, so when a borrow fails it can
identify which thread(s) are using the database.