DatabasePool allows multiple borrows of r/o db on the same thread
If a thread that currently has a read-only BorrowedDatabase calls
borrow() again, the result will now be a BorrowedDatabase on the
_same_ C4Database instance. This doesn't use up any of the other
databases in the pool, so it always succeeds. Previously, this could
cause a deadlock if all the available dbs were used up.

(This is safe because (a) it's on the same thread so there are no
concurrency issues, and (b) it's a read-only db so one caller can't
make changes that would break the other one.)

This eliminates the need for the deadlock workaround I just added
to Pusher+Revs.cc, so I've removed it.