Use DatabasePool in replicator
DBAccess now has a DatabasePool instead of using access_lock.
If constructed the normal way, it creates its own pool with capacity
2, so it has one read-only and one writeable database as before.
  It can also be passed an existing DatabasePool, which is what Edge
Server does (it has a bigger pool shared by the entire server.)
There is plumbing all the way through to c4Replicator, to allow the
client to pass in its own DatabasePool.
  The API changed slightly since it now returns BorrowedDatabase or
BorrowedCollection objects.

Lots of replicator objects were using C4Collection* pointers even
when they didn't have a lock on the DBAccess. This seemed pretty
dangerous and didn't play very well with the DatabasePool, so I
changed it so each Worker knows its CollectionSpec and uses that.