ChangesFeed: Don't observe the writeable db [CBL-6763]
Regression from the DatabasePool work: The ChangesFeed now creates its
C4DatabaseObserver on the pool's writeable db. Unfortunately this
means that during a bidirectional sync it gets notified of the changes
the Inserter is making before they're committed, because the
Inserter uses the same writeable db instance. That in turn means
when the ChangesFeed tries to read the changed documents it often
won't see them because it's reading them using a different (readonly)
db, which can't see those changes until they're committed.

It gets even worse in Edge Server, because _all_ its writes go through
that same db instance. So a continuous pusher would get notified of
changes from the REST API too early and often fail to push them.

The fix is to create the C4DatabaseObserver from one of the read-only
db instances; these won't get notified of uncommitted changes.

PS: Also improved the Replicator's "documentEnded..." log message to
include whether the doc is being pushed or pulled.