CBL-6737: Merge release/3.2 to master branch (#2231)
Includes these tickets in release/3.2:
CBL-6671: Options of FTS partial index is not fully encoded in the SQ… (#2220)
CBL-6534: Fix expiration column upgrade (#2195) (#2219)
CBL-6626: SQL++ query not return expected results (#2209)
CBL-6667: Remove LIMIT restriction in vector search (#2217)
CBL-6670: Code coverage script failing on arm64 (#2218)
CBL-6576: Implement Partial Index LiteCore API (#2215)
CBL-6651: createIndex doesn't support creating partial index with N1Q… (#2214)
CBL-6371: Unnest Query run after delete index produces obtuse error message (#2178)
CBL-6625: SQL++ query throws error unexpectedly (#2201)

* Update XCode project version to 3.2.2 (#2198)

* CBL-6625: SQL++ query throws error unexpectedly (#2201)

* CBD-6163: BSL License updates - Couchbase Lite Version 3.2 (#2207)

Change-Id: Ibc0efec8cff9eea29d1fbbfb087e931f92e09fa1

* CBL-6371: Unnest Query run after delete index produces obtuse error message (#2178)

We enhanced the error by adding to the original message, "This table is referenced by an array index, which may have been deleted."

* We also enhanced the error message of FTS index in the similar situation.

* CBL-6651: createIndex doesn't support creating partial index with N1Q… (#2214)

* CBL-6651: createIndex doesn't support creating partial index with N1QL language

Extended IndexSpec to allow the where clause when the query langugage is N1QL. We put it in IndexSpec::options. In this commit, we only do it for ValueIndex and FTSIndex.

With JSON query language, the where clause can be either in the new option or in IndexSpec::expression together with the what clause. With N1QL query language, the where clause has to be in the field of IndexSpec, and IndexSpec::expression contains only the what clause.

This commit is to get ready for C4IndexOptions

* CBL-6576: Implement Partial Index LiteCore API (#2215)

Added field "where" to C4IndexOptions and implemented it.

* CBL-6670: Code coverage script failing on arm64 (#2218)

* CBL-6667: Remove LIMIT restriction in vector search (#2217)

The removal of restriction of "LIMIT > 0" matches that in 4.0 line.

* CBL-6626: SQL++ query not return expected results (#2209)

* CBL-6534: Fix expiration column upgrade (#2195) (#2219)

* CBL-6534: Fix expiration column upgrade (#2195)

Co-authored-by: jianminzhao <76990468+jianminzhao@users.noreply.github.com>

* CBL-6671: Options of FTS partial index is not fully encoded in the SQ… (#2220)

* CBL-6671: Options of FTS partial index is not fully encoded in the SQLite table when it is registered

1. Bumped Schema version.
2. In the new version, added a column, whereClause, to table indexes.
3. Upgrade the tables in previous version to include the new column
4. IndexSpec::whereClause is registered in table "indexes" by the new column.
5. c4index_getOptions will have "where" passed in with C4IndexOptions.
6. Tests of the above.

---------

Co-authored-by: Vlad Velicu <vlad_vel@yahoo.com>
Co-authored-by: callumbirks <73551885+callumbirks@users.noreply.github.com>
Co-authored-by: Couchbase Robot <build-team@couchbase.com>