Logging: Use a separate mutex to protect sObjectMap [CBL-6787]
To avoid a deadlock on Android, we prevent the `Logging` class's
destructor from requiring the `sLogMutex`. Instead, the functions
that use `sObjectMap` will use their own mutex.

This is straightforward except for the way Logging passes a reference
to `sObjectMap` to the LogEncoder, totally breaking encapsulation.
Sigh. I fixed this the way I did on the log-observer branch, by
having the LogEncoder take the object path directly and giving the
logger a way to tell whether it needs to compute the path.