Further logging-deadlock workarounds (#2241)
* Logging: Use timed_mutex to work around deadlocks

There _should_ never be deadlocks involving the log mutex, but it's
proven very difficult to eradicate them, especially on platforms
that use GC and end up invoking LiteCore on system finalizer threads.

To mitigate the effect of a deadlock triggered by calls to log()
functions, this commit changes the bottleneck LogDomain::vlog() so it
times out after five seconds trying to acquire the mutex. The result
is that the message falls on the floor (well, it does write it to
stderr), and there's a delay, but the app does not deadlock.

* Commented out logging calls in Logging destructor overrides

These can be restored after we adopt the new logging implementation
from the feature/log-observer branch.