API: c4repl_getResponseHeaders now returns C4SliceResult [CBL-6839]
Had to change the return type because returning C4Slice is unsafe:
the implementation calls C4Replicator::getResponseHeaders which
returns alloc_slice. That heap block might be freed at any moment
after the function returns. It generally isn't, until/unless the
replicator reconnects, which is why we haven't seen bugs from this.
I discovered it when adding some more LIFETIMEBOUND annotations to
slice.hh, which then flagged this as an error.