Improve display of IPv6 addrs that wrap IPv4
An IPv4 address can be wrapped in an IPv6 address by prefixing it with
the 12 bytes {0,0,0,0,0,0,0,0,0,0,0xff,0xff}. In text form this looks
like "[::ffff:192.1.56.10]:80", not the usual "192.1.56.10:80"

The `getpeername` system call on macOS is returning this type of
address for some incoming TCP sockets. To make these easier to read
in logs, I added some code to convert them to the regular IPv4 form.