Replicator: Suppress notifications of no-op inserts
In a P2P mesh it's quite common to receive the same rev simultaneously
from multiple peers. This is causing multiple `documentsEnded`
notifications of the same revision. Additionally, the peer(s) that
sent the redundant revs think they succeeded and notify their
observers with redundant `documentsEnded` notifications too.

The database treats the extra inserts as no-ops, but that doesn't get
communicated back to the replicator. This commit remedies that.

It also adds a `noop: true` property to a `rev` message response, to
tell the peer that the revision was redundant so it doesn't tell its
observers.