Keep pulled revs from being echoed back to the peer by the Pusher
This fixes a performance issue in bidirectional continuous sync.
After the puller saves revs received from the peer, the ChangesFeed
observes those revs being added to the db and tells the pusher to
send them to the peer, which then rejects them with a 304.

The fix is to briefly track the revisions being added to the local
db, then ignore them when they're reported by the C4DatabaseObserver.