Fix Pusher bugs that omit sequences from checkpoint
I saw a failure of "Continuous Super-Fast Push" from a Windows build,
where the ending checkpoint was incorrect. Near the end of that test
it logged the active replicator checkpoint as:
    `{"local":42,"localCompleted":[0,43,48,154]}`
which is clearly wrong as it shows some sequences unable to be sent.

The test log showed the Pusher holding off on a bunch of consecutive
sequences ("...Found 0 changes up to #43" etc) because it was already
sending a revision of that doc. This was logged for sequences 43 thru
48, which is very close to the missing sequences in the checkpoint
(43-47).

From inspection, I deduced the problem to be that the code block
with comment "This doc already has a revision being sent; wait till
that one is done" stores the current rev into `iDoc->second->nextRev`
to be processed later; but if that field already holds a revision
that one gets dropped on the floor: it's obsoleted by the current
one and doesn't need to be sent, but the checkpointer is never
notified.

Adding a call to `_checkpointer.completedSequence` in that case should
fix the problem. (Also, the existing call to
`_checkpointer.addPendingSequence` is unnecessary since the Changes-
Feed already marks sequences as pending.)

Unfortunately I'm unable to reproduce the test failure locally; it
probably only occurs because the test machine is slow.

PS: Later I found another bug with the same effect, this one involving
`_conflictsIMightRetry`. Added it to the fix.