CBL-6799: Race can cause _responseTimer to be reset before stop is called
We try to avoid deleting the timer objects, _pingTimer and _responseTimer, because it's hard to synchronize their uses and deletions. Instead, We disable them, which makes the callback function a no-op function. The timers will be deleted with "this" object.

Also put ExpectingException to some tests to turn off the annoyances when running tests in XCode.