CBL-6971: Cannot access the kCBLCertAttrKeyLocality value from the certificate's attributes
In LiteCore, the corresponding constant is kC4Cert_Locality.

The component, "locality", of the subject name of the certificate works when it's used to create the certificate. When query the component of a certificate, it's shown as "L". The pleasant fact is that the abbreviated name, "L", also works to create the certificate.

We change it to "L" from "locality".

Also fixed an uninitialized error complained by Xcode analyzer.