CBL-7006: Blob changes are not sent when using delta sync in peer-to-peer replication (#2284)
The issue is ascribable to the failure of the heuristics we use to check whether a delta of a rev includes a chage of a blob. We added a situation when the only change is the content of a blob, which is reflected by a "digest" property in the delta.