Taking Jim's suggestion, we switched the focus.
Instead of trying to maintain invariant that sDefaultLogFiles be registered with only enabled levels (levels excluding kC4LogNone), we focus on the compatibility with v3.2 in separately configuring log files and log level, as manifested by two separate functions, c4log_writeToBinaryFile and c4log_setBinaryLevel.

With the former approach, we stage the result if the log level passed to c4log_writeToBinaryFile is LogNone and create sDefaultLogFiles only when enabled log level is available via c4log_setBinaryLevel. In the new approach, there can be a new state whereby there's a LogFiles observer but not wired in the log system with a enabled level. In the future for a major release, we can deprecate separate set level method, or both writeToBinary and setBinaryLevel methods.