CBL-7058: c4log_writeToBinaryFile doesn't set logging information whe… (#2287)
* CBL-7058: c4log_writeToBinaryFile doesn't set logging information when level is None
Instead of trying to maintain invariant that sDefaultLogFiles be registered with only enabled levels (levels excluding kC4LogNone), we focus on the compatibility with v3.2 in separately configuring log files and log level, as manifested by two separate functions, c4log_writeToBinaryFile and c4log_setBinaryLevel.

In this approach, there can be a new state whereby there's a LogFiles observer but not wired in the log system with a enabled level. In the future for a major release, we can deprecate separate set level method, or both writeToBinary and setBinaryLevel methods.