CBL-7099: Passive Replicator in Multipeer Replication stayed in BUSY state (#2300)
The root cause is the following. Replicator keeps the status of the pushers and pullers. They are initialized at Busy. Workers in general initializes the status either Idle or Connecting. When initialized to Idle, it may stay in Idle and since there is no change in the status, it won't have chance to update the parent Replicator.

I changed the Worker to initialize the status to Busy or Connecting. Workers cannot stay in Busy if there is no real activity.