CBL-7161: Multipeer replication fails when trying to perform merge co… (#2306)
* CBL-7161: Multipeer replication fails when trying to perform merge conflicts

The issue is caused by the function that generates the rev ID of RevTrees. In this function, the document body passed in as an argument is encoded with the SharedKeys. The actual keys are not present in "body." In lieu of them are the indexes in the SharedKeys. Therefore, documents that differ only by keys may yield identical digests which are fed to the function that generates the rev ID.

We fixed it by applying the SharedKeys to the document body before the digest is computed.

Tests that assert rev IDs are adjusted to changes in computing the digest.