CBL-7181: Collections mismatch may trigger assertion in MessageIn::respond CBL-7188: Assertion failure in LoopbackWebSocket::Driver::_close
The assertion is because that setMsgHandlerFor3_0_Client called in Replicator::handleGetCheckpoint may aleady responded with error if the passive replicagor does not have the default collection in its configuration. We fixed it by checking whether the request is already responded and avoid responding again.

CBL-7188 was encountered while working on CBL-7181