WebSocket: Fix for short heartbeat intervals
If the WS heartbeat interval is set to 10 sec or lower, the socket
wouldn't detect missing PONGs and so wouldn't close.
The reason is that every time it sends a PING it resets the 10-sec
PONG timeout, so the PONG timeout is never triggered.
Fixed by ensuring that the PONG timeout is at least one second
shorter than the PING interval.