Implemented Version::legacyVersion(revid)
Our now-standard encoding for representing a legacy tree-based revid
as a Version.

Also, C4Document::equalRevIDs() (aka revid::isEquivalentTo())
knows to treat the two forms as equivalent:
If one revID is a legacy tree-based one,
and the other revID is a Version synthesized from a legacy id,
it returns true if they represent equivalent legacy ids.