Take kSynced flag into account before RevFinder calls setDocRemoteAncestor
When RevFinder sees that a revID from the peer isn't the one stored
in the local doc for that remote, it calls DBAccess's
setDocRemoteAncestor().
However, the code that checks the local doc's remote (in VectorRecord)
wasn't taking into account the doc's kSynced flag, which tells it
that the doc's current revision is also current on remote #1.
That led to a bunch of unnecessary doc updates to set the remote rev.
Fixed that.