VectorRecord can now save with a legacy revID
Prior to this, VectorRecord::save() required the new revID to be a
Version, making it impossible to update a non-upgraded doc with a
legacy revid.

Now, if a document hasn't yet been upgraded to a version vector,
and the replicator pulls a new rev with a legacy revid,
the doc can be saved (as a VectorRecord) with that new legacy revid.

This eliminates a lot of problems with replication!