CBL-7413: Wrong error shown when using mismatch collections in Replicator
The "wrong" error is attributable to the situation when the replicator cannot find a handler for a specific message. And the situation arises when we clear the handler map as either the replicator turns to Stopped or is teminated. In either the case, we don't need to handle the message and generate the error.