Fixed bug in Version::updateClock()
If the `anyone` is false (as it is by default), the method is
supposed to skip the call to `clock.see()` if this Version is by
the local author.

Unfortunately I got the logic wrong and it ends up skipping
`clock.see()` if this Version is by _anyone but_ the local author.

That's bad because we always want to check timestamps from other
authors, because their clocks might be skewed.

Fixing this exposed some issues in unit tests where the expected
timestamp of the current rev was too low, because it didn't take
into account bumping the local clock to match existing versions.