Fixed problems syncing partially VV-upgraded database [CBL-6666, CBL-6593] (#2335)
* Implemented Version::legacyVersion(revid)

Our now-standard encoding for representing a legacy tree-based revid
as a Version.

Also, C4Document::equalRevIDs() (aka revid::isEquivalentTo())
knows to treat the two forms as equivalent:
If one revID is a legacy tree-based one,
and the other revID is a Version synthesized from a legacy id,
it returns true if they represent equivalent legacy ids.

* VectorRecord can now save with a legacy revID

Prior to this, VectorRecord::save() required the new revID to be a
Version, making it impossible to update a non-upgraded doc with a
legacy revid.

Now, if a document hasn't yet been upgraded to a version vector,
and the replicator pulls a new rev with a legacy revid,
the doc can be saved (as a VectorRecord) with that new legacy revid.

This eliminates a lot of problems with replication!

* Replicator can handle legacy revIDs in VV database [CBL-6666]

* RevFinder: in proposeChanges, check parentRev even with VVs [CBL-6666]

SG checks parentRev; we should too, at least so our unit tests will be
more indicative of what happens with SG.
(Plus, this simplifies the implementation.)

* Pusher: make shouldRetryConflictWithNewerAncestor work with VVs [CBL-6593]

It was calling C4Document::selectCommonAncestorRevision(), which
doesn't work with VectorDocuments, and is the wrong thing to check
(we want to know if receivedRevID _is_ an ancestor, not if it shares
a common ancestor.)

Added a new method C4Document::currentRevDescendsFrom() for use by
this, with implementations in both subclasses.

* Take kSynced flag into account before RevFinder calls setDocRemoteAncestor

When RevFinder sees that a revID from the peer isn't the one stored
in the local doc for that remote, it calls DBAccess's
setDocRemoteAncestor().
However, the code that checks the local doc's remote (in VectorRecord)
wasn't taking into account the doc's kSynced flag, which tells it
that the doc's current revision is also current on remote #1.
That led to a bunch of unnecessary doc updates to set the remote rev.
Fixed that.

* More tests for replicating DBs partly upgraded to VVs [CBL-6666]