Allow pushing branch-switch to server [CBL-7409]
When resolving a conflict, the convention is to always let the
non-current branch win, because that branch came from a server and we
want to preserve the server's linear history.

However, it may be impossible to do this when there are multiple
servers/remotes, because the doc's current revision may be current on
one remote while the conflict is current on another ... so merging
the conflict will switch away from one remote's branch no matter
which we let win.

This commit makes that scenario work. It allows the pusher to send a
`rev` message even when the doc's current revision is not a descendent
of the remote's current rev. (Such a rev won't have the 'noconflict'
property.)

To help out the server I added a `deleted_branch` property to the
`rev` message in this case, which contains the revision history from
a leaf (tombstone) back to the server's current rev. However, it
turns out LiteCore doesn't need this: when the passive puller saves
the revision C4Document detects it as a "server branch switch" and
updates the rev tree properly.

Sync Gateway might not handle this, though, in which case it'll need
to detect the `deleted_branch` property and act accordingly.