Updated VV conflict resolution when keeping winning rev's body
When resolving a conflict, if the merged revision has the same body
as the winning rev, generate a "trivial merge" version vector.
This doesn't have the form of a merge VV; instead it just adds the
versions from the losing VV but keeps the same current version as
the winner.