Detect redundant merges when pulling revs [CBL-7457] (#2353)
When a VV document adds an existing [pulled] revision, it now checks
if it's a redundant merge: if it has the same body and the same
merge parents (MV) as the current revision, even if the revID (CV)
is different.

In this situation it doesn't update the current local revision,
though it updates the remote's current rev as usual.