Updated VV conflict resolution when keeping winning rev's body (#2352)
* VersionVector: Added trivialMerge()

trivialMerge() merges two vectors without creating a merge version.
The resulting current version will be that of the first ("winning")
vector. This will be used when merging a conflict if the merged
body is equal to the winning rev's body.

* Updated VV conflict resolution when keeping winning rev's body

When resolving a conflict, if the merged revision has the same body
as the winning rev, generate a "trivial merge" version vector.
This doesn't have the form of a merge VV; instead it just adds the
versions from the losing VV but keeps the same current version as
the winner.

* Updated VV trivialMerge to preserve MV if possible

If the winner is a merge vector, we want to preserve that in the
result, _unless_ the loser has newer versions by any of the authors
in the winning MV. In that case we make the result a regular vector.

---------

Co-authored-by: Jianmin Zhao <jianmin.zhao@couchbase.com>