CBL-7470: Merge with release/3.3 (#2356)
* CBL-6649: Thread-safety problems with mbedTLS (#2346)

!NO_BD_GITHUB modified vendor/mbedtls/include/mbedtls/mbedtls_config.h

Actual changes are in submodule mbedtls.

* Fixed race condition calling Replicator::httpResponse() [CBL-7419] (#2343) (#2347)

...after the connection has closed.
Solution is to copy the response to the Replicator when it arrives,
instead of assuming the WebSocket is still around at the time of the
call.

Co-authored-by: Jens Alfke <jens@couchbase.com>

* Downgrade a few really noisy DB logs to 'verbose'

* When merging conflict, don't forget remotes' current revs [CBL-7409]

Resolving a conflict sometimes purges the losing branch.
But it shouldn't do that if the losing branch includes a current
remote revision, otherwise the next time the doc is pushed to that
remote we won't know the 'serverRev' to send.

* Allow pushing branch-switch to server [CBL-7409]

When resolving a conflict, the convention is to always let the
non-current branch win, because that branch came from a server and we
want to preserve the server's linear history.

However, it may be impossible to do this when there are multiple
servers/remotes, because the doc's current revision may be current on
one remote while the conflict is current on another ... so merging
the conflict will switch away from one remote's branch no matter
which we let win.

This commit makes that scenario work. It allows the pusher to send a
`rev` message even when the doc's current revision is not a descendent
of the remote's current rev. (Such a rev won't have the 'noconflict'
property.)

To help out the server I added a `deleted_branch` property to the
`rev` message in this case, which contains the revision history from
a leaf (tombstone) back to the server's current rev. However, it
turns out LiteCore doesn't need this: when the passive puller saves
the revision C4Document detects it as a "server branch switch" and
updates the rev tree properly.

Sync Gateway might not handle this, though, in which case it'll need
to detect the `deleted_branch` property and act accordingly.

* CBL-7443: Segmentation fault in Android 3.3.0-59 build while running … (#2354)

* CBL-7443: Segmentation fault in Android 3.3.0-59 build while running volume test

The stack trace points to litecore::websocket::WebSocket::validatePeerCert(). It uses a callback function that captures "this" of PeerReplicateTask. To prevent it from being called after the object is deleted, we keep a weak_ptr in WebSocket. The callback shared_ptr will be nullified in ReplicateTask::stop.

* Fixed test "Document Conflict"

---------

Co-authored-by: Jens Alfke <jens@couchbase.com>