CBL-7461: Possible racing in Replicator & WebSocket compound
Addressing a possible race condition as
- Replicator::onConnect() and
- BLIPIO::_closed()

BLIPIO::_closed() can be called from WebSocket not going through the state Connection::kClosing, but because of an error condition. It can happen as the Replicator goes from kConnecting to onConnect(). When this race occurs, Replicator::onConnect() can encouter the webSocket being nullified by BLIPIO::_closed.

We protect it by introducing Connection::webSocketSafe(). With it, if Replicator gets non-null webSocket, the the webSocket won't be deleted in BLIPIO::_closed until Replicator::onConnect() is done.