CBL-7565: Replicator stopped with Duplicate ID in version vector erro… (#2369)
* CBL-7565: Replicator stopped with Duplicate ID in version vector error when pushing a doc

The cause of this bug is the following. In LiteCore, a Version an HLV may use a special SourceID (kMeSourceID). It stands for the SourceID of whatever database it belongs to. When putting in BLIP message, it is expanded to actual SourceID. The problem is, after rounds of replication, the absolute form of SourceID may come back in the history of a HLV. This caused confustion when the local VV is compared with a VV coming from BLIP messages.

We fixed it in VersionVecWithLegacy::parse where we pass the actual SourceID to the Version constructor to turn absolute form to local form.